Fix tagfile entry not being removed from MD5 manifest files when a preserved version of a file is deleted
When a file's deleted, its corresponding tag file entries are removed from the tag files. Also, any preservation format files created of that file are also deleted. The preservation format file when deleted left behind its own MD5 manifest file tagfile entry that would result in bag integrity warnings.

+++ DataCommons/src/main/java/au/edu/anu/datacommons/storage/event/TagFilesStorageEventListener.java
@@ -50,6 +50,7 @@ import au.edu.anu.datacommons.security.service.FedoraObjectService;
 import au.edu.anu.datacommons.storage.archive.ArchiveTask;
 import au.edu.anu.datacommons.storage.archive.ArchiveTask.Operation;
 import au.edu.anu.datacommons.storage.event.StorageEvent.EventType;
+import au.edu.anu.datacommons.storage.event.StorageEventListener.EventTime;
 import au.edu.anu.datacommons.storage.event.tasks.BagCompletionTask;
 import au.edu.anu.datacommons.storage.event.tasks.ManifestMd5Task;
 import au.edu.anu.datacommons.storage.event.tasks.MetadataTask;
@@ -227,9 +228,11 @@ public class TagFilesStorageEventListener implements StorageEventListener {
 		// Delete file
 		if (event.getType().isOneOf(EventType.DELETE_FILE)) {
 			String presvRelpath = tagFilesSvc.getEntryValue(event.getPid(), PreservationMapTagFile.class, dataPrependedRelPath);
-			if (presvRelpath != null && !presvRelpath.equals("UNCONVERTIBLE")) {
-				tagFilesSvc.removeEntry(event.getPid(), PreservationMapTagFile.class, presvRelpath);
+			if (presvRelpath != null && !presvRelpath.equals("UNCONVERTIBLE") && !presvRelpath.equals("PRESERVED")) {
+				StorageEvent presvFileDelEvt = new StorageEvent(EventType.DELETE_FILE, event.getPid(), event.getProvider(), presvRelpath.replaceFirst("^data/", ""));
+				notify(EventTime.PRE, presvFileDelEvt);
 				event.getProvider().deleteFile(event.getPid(), presvRelpath.replaceFirst("^data/", ""));
+				notify(EventTime.POST, presvFileDelEvt);
 			}
 			tagFilesSvc.removeEntry(event.getPid(), PreservationMapTagFile.class, dataPrependedRelPath);
